fix: review comments

+++ internal/db/model/delegation.go
@@ -106,7 +106,7 @@ func FromEventBTCDelegationCreated(
 		log.Warn().Stringer("staking_tx_hash_hex", stakingTx.TxHash()).Msg("Staker address is empty")
 	}
 
-	if int(stakingOutputIdx) >= len(stakingTx.TxOut) {
+	if uint64(stakingOutputIdx) >= uint64(len(stakingTx.TxOut)) {
 		return nil, fmt.Errorf("staking output index %d out of range (tx has %d outputs)", stakingOutputIdx, len(stakingTx.TxOut))
 	}
 	stakingValue := btcutil.Amount(stakingTx.TxOut[stakingOutputIdx].Value)
